api/requirements.txt
azure-storage
Flask
Flask-Cors
Pillow
python-dotenv
redis
requests
numpy
